CI fix round 3: fix mce_inject.c asm constraints, splash.c float-conversion, checkpoint.c vfs_open, oci_spec.h string.h, smp.h sign-conversion, syscall.c sign-conversion
- mce_inject.c: Replace 16-operand memory asm with single-register base pointer approach to avoid Clang register pressure error; fix 'asm' vs '__asm__' for -std=c17
- splash.c: Cast float multiplication result to int
- checkpoint.c: vfs_open undeclared → declare in fs.h
- oci_spec.h: strcmp used but string.h not included
- smp.h: uint32_t→int sign conversion in smp_get_cpu_id()
- syscall.c: int→uint64_t sign conversion in copy_from_user size arg

+++ src/include/smp.h
@@ -112,7 +112,7 @@ extern int smp_cpu_count;
 /* Get current CPU ID (inline for speed) */
 static inline int smp_get_cpu_id(void) {
     struct cpu_info *info = get_cpu_info();
-    return info ? info->cpu_id : 0;
+    return info ? (int)info->cpu_id : 0;
 }
 
 /* SMP TLB shootdown: invalidate addresses on all CPUs */
